backend/requirements/development.txt
# Include base requirements
-r base.txt

# Development tools
ipython==8.18.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==20.1.0

# Code quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
pylint==3.0.3
mypy==1.7.1
django-stubs==4.2.7

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2

# Debugging
watchdog==3.0.0